Overhaul rules system
* Constant -> Index - Realistically the only thing one cares about is patching up indexes
* ConstantRegistry -> IndexRegistry - see above
* Add Modifier
* Fix GameProfile::load_components spelling
* Split nwn1/Profile.[ch]pp to make those files more manageable.
* Move selector/qualifier/etc helper functions to nwn1 namespace
* Expand kernel::Rules with functions for dealing with qualifiers, modifiers, and requirements
* Stub Situations
* Add basic draft of Effect class
* Move InternedString from kernel namespace to nw

+++ lib/nw/objects/components/CreatureStats.cpp
@@ -86,7 +86,7 @@ nlohmann::json CreatureStats::to_json() const
     return j;
 }
 
-bool CreatureStats::add_feat(int32_t feat)
+bool CreatureStats::add_feat(size_t feat)
 {
     uint16_t ft = static_cast<uint16_t>(feat);
 
@@ -105,7 +105,7 @@ const std::vector<uint16_t>& CreatureStats::feats() const noexcept
     return feats_;
 }
 
-bool CreatureStats::has_feat(int32_t feat) const noexcept
+bool CreatureStats::has_feat(size_t feat) const noexcept
 {
     uint16_t ft = static_cast<uint16_t>(feat);
