+++ lib/nw/objects/components/CreatureStats.hpp
@@ -16,9 +16,9 @@ struct CreatureStats {
     bool to_gff(GffOutputArchiveStruct& archive) const;
     nlohmann::json to_json() const;
 
-    bool add_feat(int32_t feat);
+    bool add_feat(size_t feat);
     const std::vector<uint16_t>& feats() const noexcept;
-    bool has_feat(int32_t feat) const noexcept;
+    bool has_feat(size_t feat) const noexcept;
 
     std::array<uint8_t, 6> abilities;
     std::vector<uint8_t> skills;
